+++ NEWS.md
@@ -45,6 +45,7 @@ Development
 * Fix visualization URL with hyphens in /viz ([product#229](https://github.com/CartoDB/product/issues/229))
 * Show user info in Condensed Map Card ([CartoDB/product#247](https://github.com/CartoDB/product/issues/247))
 * Show Feedback Popup ([CartoDB/product#222](https://github.com/CartoDB/product/issues/222))
+* Deprecate Google+ API and use Google Sign-In instead ([CartoDB/product#196](https://github.com/CartoDB/product/issues/196))
 * Set condensed maps view in Search page ([CartoDB/product#240](https://github.com/CartoDB/product/issues/240))
 
 4.25.0 (2019-01-28)